src/web_backend/requirements-web.txt
# ArchiFlow Web Backend Dependencies
# Install with: pip install -r src/web_backend/requirements-web.txt

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# WebSocket (Socket.IO)
python-socketio>=5.10.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Async File Operations
aiofiles>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
pytest-cov>=4.1.0

# Development
python-dotenv>=1.0.0
